Feathers terrain/cliff edges into the black void the zoomed-out camera
exposes (density vignette at map edges, contiguity-annexed black off-area
facades, local feather for thin strips). Includes the SECTOR_MAKE 64-bit
fix, the SECTOR_RECT_DIM=10 draw-coverage enlargement, the all-maps audit
tool (ARCANUM_VOID_FADE_AUDIT=1), and the "void edge fade" cfg setting
(default on).

+++ src/game/sector.h
@@ -83,6 +83,8 @@ bool sector_history_load(GameLoadInfo* load_info);
 
 #define SECTOR_X(a) ((a) & 0x3FFFFFF)
 #define SECTOR_Y(a) (((a) >> 26) & 0x3FFFFFF)
-#define SECTOR_MAKE(a, b) ((a) | ((b) << 26))
+// Must compose in 64-bit: sector Y exceeds 5 bits map-wide, so with int operands
+// `(b) << 26` overflows int32 (UB — truncates Y mod 64) and yields a garbage id.
+#define SECTOR_MAKE(a, b) ((int64_t)(a) | ((int64_t)(b) << 26))
 
 #endif /* ARCANUM_GAME_SECTOR_H_ */
